// src/app/interfaces/common/executive.interface.ts
export interface Executive {
  _id?: string;
  name?: string;
  nameBn?: string;
  title?: string;
  select: boolean;
  titleEn?: string;
  nameEn?: string;
  image?: string;
  seoImage?: string;
  url?: string;
  description?: string;
  isHtml?: boolean;
  descriptionEn?: string;
  isHtmlEn?: boolean;
  designation?: string;
  designationBn?: string;
  designationEn?: string;
  serial?: number;
  serialEn?: number;
  executiveType?: string;
  seoDescription?: string;
  pageName?: string;
  keyWord?: string;
  createdAt?: Date;
  updatedAt?: Date;
}