// src/app/interfaces/core/filter-data.ts
import {Pagination} from "./pagination";

export interface FilterData {
  filter?: any;
  filterGroup?: FilterGroup;
  pagination?: Pagination | null;
  select?: any;
  sort?: any;
}

export interface FilterGroup {
  isGroup?: boolean;
  category?: boolean;
  subCategory?: boolean;
  brand?: boolean;
}